[REV-160] refactor: 리뷰 생성 요청 dto 변경
