+++ src/main/java/com/devcourse/ReviewRanger/question/dto/request/CreateQuestionOptionRequest.java
@@ -1,8 +1,10 @@
-package com.devcourse.ReviewRanger.review.dto;
+package com.devcourse.ReviewRanger.question.dto.request;
 
 import com.devcourse.ReviewRanger.question.domain.QuestionOption;
 
-public record QuestionOptionRequest(String optionContext) {
+public record CreateQuestionOptionRequest(
+	String optionContext
+) {
 
 	public QuestionOption toEntity() {
 		return new QuestionOption(optionContext);
