+++ src/main/java/com/devcourse/ReviewRanger/review/api/ReviewController.java
@@ -4,6 +4,7 @@ import java.util.List;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -16,6 +17,7 @@ import com.devcourse.ReviewRanger.review.application.ReviewService;
 import com.devcourse.ReviewRanger.review.domain.Review;
 import com.devcourse.ReviewRanger.review.dto.request.CreateReviewRequest;
 import com.devcourse.ReviewRanger.review.dto.response.ReviewResponse;
+import com.devcourse.ReviewRanger.user.domain.UserPrincipal;
 
 @RestController
 public class ReviewController {
@@ -26,8 +28,11 @@ public class ReviewController {
 		this.reviewService = reviewService;
 	}
 
-	@PostMapping("/surveys")
-	public ResponseEntity<Boolean> createReview(@RequestBody CreateReviewRequest createReviewRequest) {
+	@PostMapping("/reviews")
+	public ResponseEntity<Boolean> createReview(
+		@RequestBody CreateReviewRequest createReviewRequest,
+		@AuthenticationPrincipal UserPrincipal user
+	) {
 		Review review = createReviewRequest.toReview();
 		review.assignRequesterId(1L);
 		List<Question> questions = createReviewRequest.toQuestions();
