--- src/main/java/com/devcourse/ReviewRanger/review/dto/ReviewRequest.java
@@ -1,13 +1,0 @@
-package com.devcourse.ReviewRanger.review.dto;
-
-import java.time.LocalDateTime;
-
-import com.devcourse.ReviewRanger.review.domain.Review;
-import com.devcourse.ReviewRanger.review.domain.ReviewType;
-
-public record ReviewRequest(String title, String description, ReviewType reviewType, LocalDateTime closedAt) {
-
-	public Review toEntity() {
-		return new Review(title, description, reviewType, closedAt);
-	}
-}
